@merlex 2025-01-28
> Спасибо за выполненное задание Хорошая работа, однако не хватает новых юнит-тестов Итого: 7 баллов из 10. Принято

Спасибо, добавил тест и поправил ошибку